+++ src/commonMain/kotlin/client/Collection.kt
@@ -29,6 +29,11 @@ import client.utils.CServiceAdapter
 import crdtlib.crdt.DeltaCRDT
 import crdtlib.crdt.DeltaCRDTFactory
 import crdtlib.utils.VersionVector
+import kotlinx.serialization.*
+import kotlinx.serialization.json.Json
+
+@Serializable
+data class CRDTJson(val _id: String)
 
 /**
 * This class represents a collection of objects.
@@ -63,10 +68,10 @@ class Collection {
     /**
      * The objects opened within this collection indexed by reference.
      */
-    internal val openedObjectsByRef: MutableMap<DeltaCRDT, Pair<CObjectUId, Boolean>> = mutableMapOf()
+    internal val openedObjectsByRef: MutableMap<DeltaCRDT, Triple<CObjectUId, Boolean, MutableSet<NotificationHandler>>> = mutableMapOf()
 
     /**
-     * Remote updates ready to be pulled
+     * Remote updates ready to be pulled.
      */
     internal val waitingPull: MutableMap<CObjectUId, DeltaCRDT> = mutableMapOf()
 
@@ -83,8 +88,16 @@ class Collection {
     }
 
     /**
-     * Pull remote updates into the current session
-     * @param type is the consistency level of the operation
+     * Get the collection ID.
+     */
+    @Name("getId")
+    fun getId() : String {
+        return this.id
+    }
+
+    /**
+     * Pull remote updates into the current session.
+     * @param type is the consistency level of the operation.
      */
     @Name("pull")
     fun pull(type: ConsistencyLevel) {
@@ -110,7 +123,7 @@ class Collection {
      * @param objectId the name of the object.
      * @param type type of the object.
      * @param readOnly is the object open in read-only mode.
-     * @param handler currently not used.
+     * @param handler handler to call when a new update is received.
      */
     @Name("open")
     fun open(objectId: String, type: String, readOnly: Boolean, handler: NotificationHandler = { _, _ -> Unit }): DeltaCRDT {
@@ -124,7 +137,12 @@ class Collection {
 
         if (obj !== null) {
             if (this.getObjectUId(obj) === null) {
-                this.openedObjectsByRef[obj] = Pair(objectUId, readOnly)
+                this.openedObjectsByRef[obj] = Triple(objectUId, readOnly, mutableSetOf(handler))
+            } else {
+                if (this.isWritable(obj) === readOnly) {
+                    throw RuntimeException("Object has been opened in different mode.")
+                }
+                this.openedObjectsByRef[obj]!!.third!!.add(handler)
             }
             return obj
         }
@@ -132,8 +150,19 @@ class Collection {
         obj = DeltaCRDTFactory.createDeltaCRDT(type, this.attachedSession.environment)
         CServiceAdapter.getObject(this.attachedSession.getDbName(), this.attachedSession.getServiceUrl(), objectUId, this)
         this.objectsById[objectUId] = obj
-        this.openedObjectsByRef[obj] = Pair(objectUId, readOnly)
+        this.openedObjectsByRef[obj] = Triple(objectUId, readOnly, mutableSetOf(handler))
         return obj
+    }
+
+    /**
+     * Send a get request for the [obj].
+     */
+    @Name("forceGet")
+    fun forceGet(obj: DeltaCRDT) {
+        val objectUId = this.getObjectUId(obj)
+        if (objectUId !== null) {
+            CServiceAdapter.getObject(this.attachedSession.getDbName(), this.attachedSession.getServiceUrl(), objectUId, this)
+        }
     }
 
     /**
@@ -150,23 +179,59 @@ class Collection {
     }
 
     /**
-     * Get the [obj] of [CObjectUID] or null if not managed by this collection
+     * Get the [obj] of [CObjectUID] or null if not managed by this collection.
      */
     internal fun getObject(objectUId: CObjectUId): DeltaCRDT? {
         return this.objectsById[objectUId]
     }
 
     /**
-     * Get the [CObjectUID] of [obj] or null if not managed by this collection
+     * Get the [CObjectUID] of [obj] or null if not managed by this collection.
      */
     internal fun getObjectUId(obj: DeltaCRDT): CObjectUId? {
         return this.openedObjectsByRef[obj]?.first
     }
 
     /**
-     * Check if [obj] is open and writable
+     * Check if [obj] is open and writable.
      */
     internal fun isWritable(obj: DeltaCRDT): Boolean {
         return this.openedObjectsByRef[obj]?.second == false
+    }
+
+    /**
+     * Get the [NotificationHandler]s of [obj] or null if not managed by this collection
+     */
+    internal fun getHandlers(obj: DeltaCRDT): MutableSet<NotificationHandler>? {
+        return this.openedObjectsByRef[obj]?.third
+    }
+
+    /**
+     * Adds a incoming update in the cache.
+     * and notify the application if a handler is set.
+     * @param objectUId UId of the crdt.
+     * @param obj new update.
+     */
+    internal fun newUpdate(objectUId: CObjectUId, obj: DeltaCRDT) {
+        this.waitingPull[objectUId] = obj
+        val crdt = this.getObject(objectUId)
+        if (crdt !== null) {
+            val handlers = this.getHandlers(crdt)
+            if (handlers !== null) {
+                for (handler in handlers) {
+                    handler(this.attachedSession.environment.getState(), objectUId)
+                }
+            }
+        }
+    }
+
+    /**
+     * Called when a new message arrives.
+     * @param message new message data.
+     */
+    internal fun newMessage(message: String) {
+        val newCRDT = Json{ignoreUnknownKeys = true}.decodeFromString<CRDTJson>(message)
+        val objectUId : CObjectUId = Json.decodeFromString<CObjectUId>(newCRDT._id)
+        this.newUpdate(objectUId, DeltaCRDT.fromJson(message))
     }
 }